Telegram-bot/requirements.txt
python-telegram-bot[socks]==13.14